Se eliminaron carpetas repetidas
